Updates to support nightly MG test automation(#1308)
This PR includes various updates to support nightly automated MG test runs, including:
* Adding a marker which nightly scripts use to run on all visible GPUs instead of a hardcoded number of GPUs, since the scripts rely on knowing the number of GPUs being used in the tests by setting the `CUDA_VISIBLE_DEVICES` env var.
  * In the nightly scripts, the marker is used like so: `pytest -m "not preset_gpu_count" ...`
* Added a `client.wait_for_workers()` call to various setups to both match the approach taken by the `MGContext` class, and to ensure workers are running. This seemed to increase reliability in the test runs. 
  * _side note: we should decide to use only the `MGContext` class or the `client_connection` pytest fixture in these tests, since they both aim to accomplish the same thing._

Authors:
  - Rick Ratzel <rratzel@nvidia.com>

Approvers:
  - null

URL: https://github.com/rapidsai/cugraph/pull/1308
