+++ python/pytest.ini
@@ -11,6 +11,7 @@ markers =
           managedmem_off: RMM managed memory disabled
           poolallocator_on: RMM pool allocator enabled
           poolallocator_off: RMM pool allocator disabled
+          preset_gpu_count: Use a hard-coded number of GPUs for specific MG tests
           ETL: benchmarks for ETL steps
           small: small datasets
           tiny: tiny datasets
